% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eva_partitions_MissAtri.R
\name{Eva_partitions_MissAtri}
\alias{Eva_partitions_MissAtri}
\title{ELVAR algorithm can predict missing nodes attribute using the other nodes attribute.}
\usage{
Eva_partitions_MissAtri(
  G,
  resolution = 1,
  threshold = 1e-07,
  alpha = 0.5,
  order = NULL,
  Vattr.name
)
}
\arguments{
\item{G}{A node-attributed igraph object}

\item{resolution}{Tune the size of the communities. Default is 1. Community size decreases with the increment of resolution value}

\item{threshold}{If change in the quality function (score) is less than the threshold, iteration loop will be terminated}

\item{alpha}{Tuning parameter that makes balance between purity and modularity during optimization}

\item{order}{Defines numerical order among the characters present within an attribute}

\item{Vattr.name}{User-select node attributes that will be considered for Eva algorithm}
}
\value{
It is a list consists of
CommunityMembers:: is a vector where names are the nodes and the values are the communities belong to the nodes,
ModularityVal:: Modularity value, and
PurityVal:: Purity value
}
\description{
The code is an extended form of Eva_partitions() function.
It infers missing attribute of nodes from a partially node-attributed graph.
Missing attribute of nodes are assigned as "MA".
}
