% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LouvainRND.R
\name{LouvainRND}
\alias{LouvainRND}
\title{Louvain algorithm detects communities from a graph network}
\usage{
LouvainRND(G, resolution = 1, threshold = 1e-07)
}
\arguments{
\item{G}{An igraph object}

\item{resolution}{Tune the size of the communities. Default is 1. Community size decreases with the increment of resolution value}

\item{threshold}{If change in the quality function (score) is less than the threshold, iteration loop will be terminated}
}
\value{
It is a list consists of
CommunityMembers:: is a vector where names are the nodes and the values are the communities belong to the nodes, and
ModularityVal:: Modularity value
}
\description{
This function optimizes (maximize) the modularity function for identifying communities from a graph network
following Louvain-like method. Louvain implementation in igraph R package gives a unique network partition
from a graph network on every run due to sequential selection of node for local moving during optimization.
But this Louvain implementation may generate slightly different network partitions from a graph network on
every run because we considered random (nonsequential) selection of node for local moving during optimization.
}
