% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessEVA.R
\name{ProcessEVA}
\alias{ProcessEVA}
\title{Processes the output of the EVA clustering algorithm}
\usage{
ProcessEVA(eva.o, seu.o, attrName, sigth = NULL)
}
\arguments{
\item{eva.o}{The output from running the EVA-clustering algorithm.}

\item{seu.o}{The Seurat scRNA-Seq data object containing the metacell information.}

\item{attrName}{The name of the attribute is used for testing enrichment of clusters. Must be one of the names
in the metadata slot of the Seurat object.}

\item{sigth}{The significance threshold on the binomial-test P-values for declaring enrichment.
If not specified (default), the function will use a Bonferroni correction.}
}
\value{
A list with four elements

pvalEnr
The enrichment P-value matrix defined over clusters (rows) and attribute values (columns).

sigth
The significance threshold used to declare enriched clusters.

sigClust
A list of enriched cluster identifiers, each list entry representing one attribute value.

cellsMrg
A list of cell indices indicating which cells of a given attribute value are in enriched clusters
for that attribute value. Each entry in the list represents an attribute value.
}
\description{
This function takes as input the result of an EVA-clustering plus the metacell information from the Seurat
scRNA-Seq data object, to identify for each attribute value the communities enriched for cells with that
attribute value. This is done using a binomial test. Once the enriched clusters for each attribute value
are found, cells from these clusters and with the given attribute value are merged.
}
\references{
Alok K Maity, Andrew E Teschendorff
\emph{Cell-Attribute aware community detection improves differential abundance testing from single-cell RNA-Seq data}
Submitted.
}
