% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelOptAlpha.R
\name{SelOptAlpha}
\alias{SelOptAlpha}
\title{Selects optimal purity parameter value}
\usage{
SelOptAlpha(nC.m,nc0,pur.m,mod.m,alpha.v=seq(0.1,0.9,0.1),thP=0.75,thM=0.75,qC=0.95,thC=1.5)
}
\arguments{
\item{nC.m}{A matrix of inferred cluster numbers, with rows labeling runs and columns labeling the alpha-values.}

\item{nc0}{A scalar giving the number of inferred clusters using the Louvain algorithm.}

\item{pur.m}{A matrix of estimated purity values, with rows labeling the runs and columns labeling the alpha-values.}

\item{mod.m}{A matrix of estimated modularity values, with rows
  labeling the runs and columns labeling the alpha-values.}

\item{alpha.v}{A vector of purity parameter values. Should be an
  increasing sequence between 0 and 1, excluding these values as these
  are not interesting. By default, this is a sequence from 0.1 to
  0.9. We don't advise changing the extreme values, but user can change
  the granularity or step-size.}

\item{thP}{A scalar between 0 and 1, which quantifies by how much purity
  at the optimal alpha-value is allowed to deviate from the maximum
  attained over the whole range of alpha-values considered. By default
  this value is 0.75. A value of 1 would mean no deviation from the
  maximum.}

\item{thM}{A scalar between 0 and 1, which quantifies by how much the
  modularity at the optimal alpha-value is allowed to deviate from the
  maximum attained over the whole range of alpha-values considered. By
  default, this values is 0.75. A value of 1 would mean no deviation
  from the maximum.}

\item{qC}{A scalar between 0 and 1 labeling the quantile of the EVA
  cluster number distribution to consider. By default, this value is
  0.95, which means that the upper 95 percent quantile of the EVA
  cluster number distribution is used when comparing to \code{nc0}.}

\item{thC}{A scalar quantifying by how much higher the EVA cluster
  number obtained via \code{qC} should be in relation to \code{nc0}. By
  default, this value is 1.5, which means that we require the upper 95
  percent quantile of the EVA cluster number distribution to be at least
  1.5 higher than \code{nc0}.}

}
\value{
A list with three elements

aM
The largest allowed alpha-value satisfying the modularity constraint.

aP
The smallest allowed alpha-value satisfying the purity constraint.

aC
The smallest allowed alpha-value satisfying the cluster number constraint.
}

\description{
 The aim of this function is to facilitate choosing an optimal purity parameter (alpha) value. As input it requires the number of inferred clusters, the purity and modularity values for a range of different alpha values and for a reasonably large number of runs (typically we recommend 100 runs), as well as the number of clusters inferred with the ordinary Louvain algorithm. If a non-sequential stochastic version of Louvain is used, then this could be an average value over multiple runs.
}
\references{
Alok K Maity, Andrew E Teschendorff
\emph{Cell-Attribute aware community detection improves differential
 abundance testing from single-cell RNA-Seq data}. Submitted.
}
