% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoDA.R
\name{DoDA}
\alias{DoDA}
\title{Performs DA-analysis}
\usage{
DoDA(seu.o, sigcl.o, varDA, varREP)
}
\arguments{
\item{seu.o}{The Seurat scRNA-Seq data object containing the metacell information.}

\item{sigcl.o}{The output from running the \code{ProcessEVA} function.}

\item{varDA}{A character specifying the name of the cell-state for which DA is being
tested for.}

\item{varREP}{A character specifying the name of replicate ID variable.}
}
\value{
A list with two elements

stat
A matrix of statistics (Wald and LRT) from the negative binomial regression, with each row
representing a cell-state and columns labeling the z-statistic and P-value corresponding to
the statistical tests.

df
The data frame used for the negative binomial regression. Each row labels one
attribute-replicate pair. Columns label the attribute value, the two states for
which DA is to be tested across the attribute, and the normalization factor.
}
\description{
This function performs differential abundance (DA) testing of a cell-state variable specified
by \code{varDA} across the cell-attribute used in the previous EVA-clustering step. Replicate
information is specified via the \code{varREP} variable. The cell annotations must be present in
the metadata slots of the Seurat scRNA-Seq data object \code{seu.o}. The other inputs to this function
include the output of one EVA-clustering run, as well as the output from \code{ProcessEVA}. The function
then returns the negative binomial regression statistics (Wald and LRT) to determine if the cell-state fractions vary
significantly as a function of attribute value.
}
\references{
Alok K Maity, Andrew E Teschendorff
\emph{Cell-Attribute aware community detection improves differential abundance testing from single-cell RNA-Seq data}
Submitted.
}
